# swaggerapi.yaml
openapi: 3.0.0

info: 
  title: The TreeGarden Project
  description: This is the document of TreeGarden Function
  version : 1.0.0
  contact: 
    email: tranductai@gspit.net
  liscence: 
    name: TreeGarden 1.0
    url : "http://TreeGarden.org/"
  server: [localhost:3000/]


paths:
#get all plants
  /plants:
    get:
      
      tags:
        - plants
      summary: these are plants
      description: these are all the plant we have 
      security:
        - bearerAuth: []
      parameters:
      - $ref: "#/parameters/page"
      - $ref: "#/parameters/limit"
      responses: 
        200:
          description : success
  
# #get plant by page and limit query
#   /plants/?:
#     get:
#       tags:
#         - plants
#       summary: get plant by query
#       description: enter page number or/and limit
#       parameters:
#         required: true
#         content:
#           - page: page
#             in: path
#             schema: 
#               type: number
#           - limit: limit
#             in: path
#             schema:
#               type: number
#       responses:
#         200:
#           description : success


#get all kinds
  /kinds:
    get:
      tags:
        - kinds
      summary: these are kinds
      description: these are all the kinds we have 
      responses: 
        200:
          description : success
  /kinds/{id}:
    get:
      tags:
        - kinds
      summary: get kinds by id kinds
      description: these are all kinds we find by id 
      parameters:
        - name : id
          in: path
          required : true 
          schema:
            type: string 
      responses:
        200:
          description: find success
          schema: 
            $ref: '#/components/schemas/Kinds'
    delete:
      tags:
        - kinds
      summary: delete kinds by id kinds
      description: these are all kinds we delete by id 
      parameters:
        - name : id
          in: path
          required : true 
          schema:
            type: string 
      responses:
        200:
          description: delete success
          schema: 
            $ref: '#/components/schemas/Kinds'
    put:
      tags:
        - kinds
      summary: changes kinds by id kinds
      description: these are all kinds we changes by id 
      parameters:
        - name : id
          in: path
          required : true 
          schema:
            type: string 
      requestBody:
        required: true 
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Kinds'
      responses:
        200:
          description: delete success
          schema: 
            $ref: '#/components/schemas/Kinds'
  /kinds/post:
    post:
      tags:
        - kinds
      summary : post the kinds
      produces:
        - applcation/json
      requestBody:
        required: true 
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Kinds'
      responses:
          200:
            description: post plant success!
            schema: 
              $ref: '#/components/schemas/Kinds'
  
  # /kinds/sort:

#post data
  /plants/:
    post:
      tags:
        - plants
      summary: post the plant
      produces:
        - applcation/json
      requestBody:
        required: true 
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Plants'
      responses:
          200:
            description: post plant success!
            schema: 
            type: object
            properties: 
              $ref: '#/components/schemas/Plants'


                
#get plant by id
  /plants/{id}:
    get:
      tags:
        - plants
      summary: plant by id
      description: find  a plant by id
      parameters:
        - name : id
          in: path
          required : true 
          schema:
            type: string 
      responses: 
        200:
          description: success 
          schema: 
            type: object
            properties: 
              $ref: '#/components/schemas/Plants'

    put:
      tags:
        - plants
      summary: plant by id
      parameters:
        - name : id
          in: path
          required : true 
          schema:
            type: string  
      requestBody:
        required: true 
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Plants'
      responses:  
        200:
          description: success 
          schema: 
            type: object
            properties: 
              $ref: '#/components/schemas/Plants'
    delete:
      tags:
        - plants
      summary: Delete plant
      description: This can only be done once
      operationId: deletePlant
      parameters:
        - name: id
          in: path
          description: The id that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        200:
          description: success!


  /user/all:
    get:
      tags:
        - User  
      summary: these are user
      description: these are all the user account we have in system
      parameters:
      - $ref: "#/parameters/page"
      - $ref: "#/parameters/limit"
      responses:  
        200:
          description : success
  /user/register:
    post:
      tags:
        - User
      summary: Register User
      description: Create a user in database 
      requestBody:
        required: true 
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/login'
      responses:
        200:
          description: Register success!
  


  /user/login:
    post:
      tags:
      - User
      summary: login check
      description: checkin from client
      securityschema:
      - $ref: '#/components/securityschema'
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/login'
      responses:
        200:
          description: success
  /user/{id}:
    put:
      tags:
        - User
      summary: changes user informations
      description: these are all informations we changes by id 
      parameters:
        - name : id
          in: path
          required : true 
          schema:
            type: string 
      requestBody:
        required: true 
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        200:
          description: update success
          schema: 
            $ref: '#/components/schemas/User'
    delete:
      tags:
        - User
      summary: Delete user
      description: This can only be done once
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          description: The id that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        200:
          description: success!
          schema: 
            $ref: '#/components/schemas/User'

  /cart/addcart/{id}:
    put:
      tags: 
        - Cart
      summary: cart by id
      description: find  a cart with id user follow
      parameters:
        - name: id
          in: path
          description: The id that needs to be find
          required: true
          schema:
            type: string
      requestBody:
        required: true 
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Cart'
      responses:
        200:
          description: success!
          schema: 
            $ref: '#/components/schemas/Cart'
  /cart/{id}:
    get:
      tags:
        - Cart
      summary: cart by id
      description: get present cart with id cart follow
      parameters:
        - name: id
          in: path
          description: The id that needs to be find
          required: true
          schema:
            type: string
      responses:
        200:
          description: this is the present cart by your id 


    

components:
  schemas:
    Plants:
      type: object
      properties: 
        namePlant:
          type: string 
        kind:
          type: object
        quantity: 
          type: number
        price:
          type: number
        image: 
          type: string
        discount:
          type: number

    Kinds :
      type: object
      properties:
        nameKind:
          type: string 
        size: 
          type: string
        tempt:
          type: string
        hight: 
          type: number
        often:
          type: string 
        list:
          type : string
    User:
      type: object
      properties:
        nameUser:
          type: string
        Phone:
          type: number
        password:
          type : string
        role:
          type: string
        history:
          type: array
    login:
      type: object
      properties:
        nameUser:
          type: string
        password:
          type: string
    Cart:
      type: object 
      properties:
        plantid: 
          type: string
        quantity:
          type: number

securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: Token-based authorization

definitions:
 PaginationResponse:
    type: object
    properties:
      totalrecords:
        type: number
      displayrecords:
         type: number
    xml:
       name: PaginationResponsedata


parameters:
    limit:
      name: limit
      in: query
      description: Number of records to return
      type: number
    page:
      name: page
      in: query
      description: Start index for paging
      type: number